new migration file

+++ src/Views/Generators/migration.blade.php
@@ -23,8 +23,6 @@ class CreateFortressSetupTables extends Migration
             $table->string('name')->unique();
             $table->string('display_name')->nullable();
             $table->string('description')->nullable();
-            $table->integer('{{$organizationForeignKey}}')->unsigned()->nullable();
-            $table->foreign('{{$organizationForeignKey}}')->references('id')->on('{{$organizationsTable}}')->onUpdate('cascade')->onDelete('cascade');
             $table->timestamps();
         });
         // Create table for storing permissions
@@ -33,8 +31,6 @@ class CreateFortressSetupTables extends Migration
             $table->string('name')->unique();
             $table->string('display_name')->nullable();
             $table->string('description')->nullable();
-            $table->integer('{{$organizationForeignKey}}')->unsigned()->nullable();
-            $table->foreign('{{$organizationForeignKey}}')->references('id')->on('{{$organizationsTable}}')->onUpdate('cascade')->onDelete('cascade');
             $table->timestamps();
         });
         //Create table for storing permission_role
@@ -45,43 +41,28 @@ class CreateFortressSetupTables extends Migration
             $table->integer('{{$roleForeignKey}}')->unsigned();
             $table->foreign('{{$roleForeignKey}}')->references('id')->on('{{$rolesTable}}')->onUpdate('cascade')->onDelete('cascade');
         });
-        //Create table for storing role_user
-        Schema::create('{{$roleUserTable}}', function (Blueprint $table) {
+        // Create table for storing the persona
+        if (!Schema::hasTable('{{$personaTable}}')) {
+            Schema::create('{{$personaTable}}', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('{{$userForeignKey}}')->unsigned();
+                $table->foreign('{{$userForeignKey}}')->references('id')->on('{{$usersTable}}')->onUpdate('cascade')->onDelete('cascade');
+                $table->morphs('{{$personaTable}}');
+            });
+        };
+        // Create table for storing persona_role
+        Schema::create('{{$personaRoleTable}}', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('{{$roleForeignKey}}')->unsigned();
-            $table->foreign('{{$roleForeignKey}}')->references('id')->on('{{$rolesTable}}')->onUpdate('cascade')->onDelete('cascade');
-            $table->integer('{{$userForeignKey}}')->unsigned();
-            $table->foreign('{{$userForeignKey}}')->references('id')->on('{{$usersTable}}')->onUpdate('cascade')->onDelete('cascade');
-        });
-        //Create table for storing permission_user
-        Schema::create('{{$permissionUserTable}}', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('{{$permissionForeignKey}}')->unsigned();
-            $table->foreign('{{$permissionForeignKey}}')->references('id')->on('{{$permissionsTable}}')->onUpdate('cascade')->onDelete('cascade');
-            $table->integer('{{$userForeignKey}}')->unsigned();
-            $table->foreign('{{$userForeignKey}}')->references('id')->on('{{$usersTable}}')->onUpdate('cascade')->onDelete('cascade');
-        });
-        // Create table for storing organization_user
-        Schema::create('{{$organizationUserTable}}', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('{{$userForeignKey}}')->unsigned();
-            $table->foreign('{{$userForeignKey}}')->references('id')->on('{{$usersTable}}')->onUpdate('cascade')->onDelete('cascade');
-            $table->integer('{{$organizationForeignKey}}')->unsigned();
-            $table->foreign('{{$organizationForeignKey}}')->references('id')->on('{{$organizationsTable}}')->onUpdate('cascade')->onDelete('cascade');
-        });
-        // Create table for storing organization_user_role
-        Schema::create('{{$organizationUserRoleTable}}', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('{{$organizationUserForeignKey}}')->unsigned();
-            $table->foreign('{{$organizationUserForeignKey}}')->references('id')->on('{{$organizationUserTable}}')->onUpdate('cascade')->onDelete('cascade');
+            $table->integer('{{$personaForeignKey}}')->unsigned();
+            $table->foreign('{{$personaForeignKey}}')->references('id')->on('{{$personaTable}}')->onUpdate('cascade')->onDelete('cascade');
             $table->integer('{{$roleForeignKey}}')->unsigned();
             $table->foreign('{{$roleForeignKey}}')->references('id')->on('{{$rolesTable}}')->onUpdate('cascade')->onDelete('cascade');
         });
-        // Create table for storing organization_user_permission
-        Schema::create('{{$organizationUserPermissionTable}}', function (Blueprint $table) {
+        // Create table for storing persona_permission
+        Schema::create('{{$personaPermissionTable}}', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('{{$organizationUserForeignKey}}')->unsigned();
-            $table->foreign('{{$organizationUserForeignKey}}')->references('id')->on('{{$organizationUserTable}}');
+            $table->integer('{{$personaForeignKey}}')->unsigned();
+            $table->foreign('{{$personaForeignKey}}')->references('id')->on('{{$personaTable}}');
             $table->integer('{{$permissionForeignKey}}')->unsigned();
             $table->foreign('{{$permissionForeignKey}}')->references('id')->on('{{$permissionsTable}}')->onUpdate('cascade')->onDelete('cascade');
         });
@@ -95,11 +76,9 @@ class CreateFortressSetupTables extends Migration
      */
     public function down()
     {
-        Schema::drop('{{$organizationUserPermissionTable}}');
-        Schema::drop('{{$organizationUserRoleTable}}');
-        Schema::drop('{{$organizationUserTable}}');
-        Schema::drop('{{$permissionUserTable}}');
-        Schema::drop('{{$roleUserTable}}');
+        Schema::drop('{{$personaPermissionTable}}');
+        Schema::drop('{{$personaRoleTable}}');
+        Schema::drop('{{$personaTable}}');
         Schema::drop('{{$permissionRoleTable}}');
         Schema::drop('{{$permissionsTable}}');
         Schema::drop('{{$rolesTable}}');
